test/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.6.2.post1
asgiref==3.8.1
asn1crypto==1.5.1
astrapy==1.5.2
asyncpg==0.30.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
azure-core==1.32.0
azure-storage-blob==12.24.0
babel==2.16.0
beautifulsoup4==4.12.3
boto3==1.35.71
botocore==1.35.71
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.8
cobble==0.1.4
cohere==5.12.0
constantly==23.10.4
courlan==1.3.2
cryptography==44.0.0
cssselect==1.2.0
daphne==4.1.2
databases==0.9.0
dateparser==1.2.0
defusedxml==0.7.1
Deprecated==1.2.15
deprecation==2.1.0
diffusers==0.32.2
distro==1.9.0
Django==5.1.4
dnspython==2.7.0
docker==7.1.0
duckduckgo_search==7.3.0
e2b==1.0.5
e2b-code-interpreter==1.0.3
elevenlabs==1.13.0
email_validator==2.2.0
et_xmlfile==2.0.0
Events==0.5
exa_py==1.7.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastavro==1.9.7
fastcore==1.7.20
fastlite==0.0.13
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
griptape==1.4.2
grpcio==1.68.0
grpcio-status==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmldate==1.9.2
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
iniconfig==2.0.0
ipaddress==1.0.23
isodate==0.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jusText==3.0.1
lxml==5.3.0
lxml_html_clean==0.4.1
mail-parser==4.1.2
mammoth==1.9.0
markdown-it-py==3.0.0
markdownify==0.14.1
markitdown==0.0.1a3
MarkupSafe==3.0.2
marqo==3.10.0
marshmallow==3.23.1
marshmallow-enum==1.5.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
ndg-httpsclient==0.5.1
numexpr==2.10.2
numpy==2.2.1
oauthlib==3.2.2
ollama==0.4.7
openai==1.65.2
openpyxl==3.1.5
opensearch-py==2.7.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-threading==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parsel==1.9.1
pathvalidate==3.2.3
pdfminer.six==20240706
pgvector==0.3.6
pillow==11.0.0
pinecone-client==3.2.2
platformdirs==4.3.6
playwright==1.49.0
pluggy==1.5.0
portalocker==2.10.1
primp==0.11.0
prompt-toolkit==3.0.36
propcache==0.2.0
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.10
puremagic==1.28
pusher==3.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
pypdf==5.1.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.0
python-multipart==0.0.18
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.1
questionary==2.0.1
queuelib==1.7.0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
schema==0.7.7
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.3
snowflake-sqlalchemy==1.7.0
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.13.0
SQLAlchemy==2.0.36
sqlite-minutils==4.0.2
sqlparse==0.5.3
starlette==0.41.3
stringcase==1.2.0
tavily-python==0.5.0
tenacity==8.5.0
tiktoken==0.9.0
tld==0.13
tldextract==5.1.3
tokenizers==0.20.3
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.46.3
Twisted==24.11.0
txaio==23.1.1
typer==0.14.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
uuid6==2024.7.10
uvicorn==0.32.1
uvloop==0.21.0
voyageai==0.2.4
w3lib==2.2.1
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wrapt==1.17.0
XlsxWriter==3.2.0
yarl==1.18.0
youtube-transcript-api==0.6.3
zipp==3.21.0
zope.interface==7.2
